[Feature] [checked-permission] New component! AndesTimePicker (#348)
* dropdown form floatingmenu variant

implementation of the FLOATINGMENU variant for the AndesDropDownForm

* tests for new variant

tests for new variant

* changes for code review by tito

changes for code review by tito

* changes in menu creation

the responsibility to create and implement the menus (floating menu or bottom sheet dialog) is now part of the menu interface.

* changes in tests for the config factory

the menuType assert now asserts the menuType class in the attrs is the same as the menuType class in the config

* merge master into branch

merge master into branch

* changes by code review by tito

changes by code review by tito

* struct component

* created a11y delegate

an a11yDelegate is created and assigned to the view. Also, the navigation of the component is corrected

* tests for a11yDelegate

tests for a11yDelegate

* bug fixes in timepicker

deleted placeholder setter not needed (placeholder should always be "00:00"), deleted currentTime setter not needed. Also is corrected the attrs class without this two variables

* updates in demoapp

created dynamic page for the timepicker

* timepicker showcase static page

timepicker showcase static page implementation

* Improvements

* changes in dropdown a11y to comply with timepicker a11y

changes in the dropdown role to allow the internal textfield to be setted with a READONLY value. This is needed for the READONLY state in the timepicker

* test setReadOnly() method

test setReadOnly() method

* changes in timepicker showcase

added a textfield to set a new currentTime value in the timepicker dynamic page of the demoapp

* update changelog

update changelog

* Add test

* changes in demoapp

Changed the exception type caught when trying to set an incorrect value in the timepicker

* changes for linter

changes for linter

* changes for linter

changes for linter

* changes in timepicker config

changes in timepicker config

* changes for code reviey by cris

changes for code reviey by cris

* changes for code review by cris

changes for code review by cris

* Add ttest interfaces

* changes in config class

changes in config class. Also, a refactor typo in package name is included

* tests for timepicker

tests for timepicker

* test for dropdown a11y

needed to comply with the a11y of the timepicker

* updates in timepicker showcase

changes in the demoapp for the static page of the timepicker

* changes in code review by tito

changes in code review by tito

* fixes for setupDropdown internal method

fixes for setupDropdown internal method

* changes for code review by tito

changes for code review by tito

* timepicker docs

created AndesTimePicker docs

* changes in interface naming

changed the interface name for the time selected event. Also fixed some issues for the lint check

* changes for code review by tito

changes for code review by tito

* changes for code review by tito

changes for code review by tito

* fix linter

fix linter

* changes in clickListeners assignation

changes in clickListeners assignation

Co-authored-by: Martín Pérez <jose.martin@mercadolibre.com>